% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_test.R
\docType{data}
\name{ct_test}
\alias{ct_test}
\title{Example dataset}
\format{
A data frame with columns:
\describe{
\item{ID}{Subject ID}
\item{trt}{Treatment group}
\item{x}{Ct values for the housekeeping gene}
\item{y}{Ct values for target gene y}
\item{z}{Ct values for target gene z}
}
}
\source{
Data created by the author.
}
\usage{
ct_test
}
\description{
A dataset containing Ct values of 3 genes for  9 samples which belong to 3 treatment groups
}
\details{
A data set containing Ct values for 3 genes, with 3 replicted in each group
}
\keyword{datasets}
