% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FC.R
\name{plot_FC}
\alias{plot_FC}
\title{Bar Graph for Gene Fold Change}
\usage{
plot_FC(df.process)
}
\arguments{
\item{df.process}{The output data created from the \code{process.data} function, including the fold change

\describe{
\item{trt}{A factor column indicating treatment groups.}
\item{Fold Change Columns}{Fold change values for each target gene (e.g., \verb{FC.<gene>}).}
}}
}
\description{
Plot Fold Change for Target Genes
}
\details{
We want to use ggplot to graph the fold change for each of the target genes.
}
\author{
Xia Zhao
}
