% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFC.R
\name{PlotFC}
\alias{PlotFC}
\title{Bar Graph for Gene Fold Change}
\usage{
PlotFC(df.process)
}
\arguments{
\item{df.process}{The output data created from the \code{ProcessData} function, including the fold change

\describe{
\item{trt}{A factor column indicating treatment groups.}
\item{Fold Change Columns}{Fold change values for each target gene (e.g., \verb{FC.<gene>}).}
}}
}
\description{
Plot Fold Change for Target Genes
}
\details{
We want to use ggplot to graph the fold change for each of the target genes.
}
\examples{
# Example usage with internal CSV data bundled with the package
# Step 1: Load the data and process it to get df.process
file_path <- system.file("extdata", "ct_test.csv", package = "ddFoldChange")
output_file <- "processed_ct_test.csv"
df.process<- ProcessData(file.path = file_path, replicate = 3, output.file = output_file)
head(df.process)

# Step 2: Plot the fold change for target genes using the processed data
PlotFC(df.process)
}
\author{
Xia Zhao
}
