% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.data.R
\name{process.data}
\alias{process.data}
\title{Gene fold change calculation}
\usage{
process.data(file.path, replicate, output.file)
}
\arguments{
\item{file.path}{Path to the input CSV file}

\item{replicate}{Number of replicates for each group}

\item{output.file}{Created Output file for fold change plotting in the next function
\describe{
\item{means.internal}{Mean Ct values of the housekeeping gene for each group}
\item{dCt.col.stat}{The column corresponding to the first target gene}
\item{columns.to.normalize}{The columns corresponding to all target genes}
\item{baseline.means}{The mean delta CT value for the control group for each target gene}
\item{new.col.name}{Create new columns for delta CT, delta CT, and fold change in each loop}
}}
}
\value{
A dataframe containing original Ct value, delta CT value, delta delta CT value and fold changes
\describe{
\item{Ct values}{Original Ct values from the input file.}
\item{dCt}{Delta Ct values for each target gene.}
\item{ddCt}{Delta delta Ct values for each target gene.}
\item{FC}{Calculated fold changes for each target gene.}
}
}
\description{
Process the input Data for Fold Change Calculations
}
\details{
We want to perform normalization of Ct values for the target genes and calculate fold-change calculations relative to the control group
}
\author{
Xia Zhao
}
